refactor: move keyboard logic to keyboardUtils
